requirements.txt
evdev
pyserial
